pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "social-media-optimizer"
version = "0.1.0"
description = "RL Environment for multi-brand social media agency optimization — post timing, content mix, and budget allocation"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "fastapi>=0.115.0",
    "pydantic>=2.0.0",
    "uvicorn>=0.24.0",
    "requests>=2.31.0",
    "openai>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "social_media_optimizer.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["social_media_optimizer", "social_media_optimizer.server"]
package-dir = { "social_media_optimizer" = ".", "social_media_optimizer.server" = "server" }
